+++ domain/upload/src/main/java/com/record/upload/repository/UploadRepository.kt
@@ -7,5 +7,6 @@ import java.io.File
 interface UploadRepository {
     suspend fun getPresignedUrl(): Result<UploadInfo>
     suspend fun uploadRecord(videoInfo: VideoInfo): Result<Unit>
-    suspend fun uploadVideoToS3Bucket(url: String, file: File): Result<Unit>
+    suspend fun uploadVideoToS3Bucket(url: String, file: File): Result<String>
+
 }
