+++ feature/upload/src/main/java/com/record/upload/UploadViewModel.kt
@@ -8,7 +8,6 @@ import com.record.designsystem.component.snackbar.SnackBarType
 import com.record.keyword.repository.KeywordRepository
 import com.record.ui.base.BaseViewModel
 import com.record.upload.extension.GalleryVideo
-import com.record.upload.extension.uploadFileToS3PresignedUrl
 import com.record.upload.extension.uploadFileToS3ThumbnailPresignedUrl
 import com.record.upload.model.VideoInfo
 import com.record.upload.repository.UploadRepository
@@ -46,23 +45,19 @@ class UploadViewModel @Inject constructor(
         viewModelScope.launch(Dispatchers.IO) {
             var a = ""
             var b = ""
-            uploadFileToS3PresignedUrl(
+            uploadRepository.uploadVideoToS3Bucket(
                 uiState.value.bucketUrl,
-                uiState.value.thumbnailUrl,
                 file,
-            ) { success, message ->
-                println(message)
-                a = removeQueryParameters(message)
-                if (success) {
-                    uploadFileToS3ThumbnailPresignedUrl(
-                        context,
-                        uiState.value.thumbnailUrl,
-                        file,
-                    ) { success, message ->
-                        println(message)
-                        b = removeQueryParameters(message)
-                        uploadRecord(a, b)
-                    }
+            ).onSuccess {
+                a = removeQueryParameters(it)
+                uploadFileToS3ThumbnailPresignedUrl(
+                    context,
+                    uiState.value.thumbnailUrl,
+                    file,
+                ) { success, message ->
+                    println(message)
+                    b = removeQueryParameters(message)
+                    uploadRecord(a, b)
                 }
             }
         }
